Implement the app topic modal to the topics page
