+++ frontend/package.json
@@ -25,11 +25,11 @@
     "@fortawesome/fontawesome-free": "^5.13.1",
     "bootstrap": "^4.5.0",
     "concurrently": "^5.2.0",
+    "dev-server": "^0.2.0",
     "faker": "^4.1.0",
     "fuzzysort": "^1.2.1",
     "json-server": "^0.16.1",
     "rxjs": "~6.5.4",
-    "tslib": "^1.10.0",
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
@@ -50,6 +50,7 @@
     "karma-jasmine-html-reporter": "^1.4.2",
     "protractor": "~5.4.3",
     "ts-node": "~8.3.0",
+    "tslib": "^2.4.0",
     "tslint": "~6.1.0",
     "typescript": "~3.8.3"
   }
