+++ frontend/src/app/topic-page/topic-page.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router} from '@angular/router';
 import { HttpClient } from '@angular/common/http';
 import { TopicResult } from './TopicResult';
+import { ModalService } from '../_modal';
 
 @Component({
   selector: 'app-topic-page',
@@ -15,8 +16,10 @@ export class TopicPageComponent implements OnInit {
   id: string;
   sub;
   results: TopicResult;
+  topicTitle: string;
+  topicDescription: string;
 
-  constructor(private _Activatedroute:ActivatedRoute, private _router:Router, private http: HttpClient) { 
+  constructor(private _Activatedroute:ActivatedRoute, private _router:Router, private http: HttpClient, private modalService: ModalService) { 
     this.sub=this._Activatedroute.paramMap.subscribe(params => { 
       console.log(params);
        this.id = params.get('id'); 
@@ -39,8 +42,27 @@ export class TopicPageComponent implements OnInit {
     console.log("test");
   }
 
-  addTopic(){
+  addTopic(classID: string){
+    var URL = 'https://zipit-backend.herokuapp.com/add/topic';
+    var body = {
+      "title": this.topicTitle,
+      "description": this.topicDescription,
+      "class_id": classID
+    };
+    this.closeModal('addTopic');
+    this.http.post(URL, body);
+  }
 
+  closeModal(id: string){
+    this.modalService.close(id);
+    this.topicTitle = "";
+    this.topicDescription = "";
+  }
+
+  openModal(id: string){
+    this.modalService.open(id);
+    this.topicTitle = "";
+    this.topicDescription = "";
   }
 
 }
